requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
Pypdf2
chromadb